refactor: fix clippy warnings (#2515)

+++ packages/wasm-dpp/src/identity/state_transition/identity_credit_withdrawal_transition/transition.rs
@@ -138,7 +138,7 @@ impl IdentityCreditWithdrawalTransitionWasm {
 
     #[wasm_bindgen(js_name=getUserFeeIncrease)]
     pub fn get_user_fee_increase(&self) -> u16 {
-        self.0.user_fee_increase() as u16
+        self.0.user_fee_increase()
     }
 
     #[wasm_bindgen(js_name=setUserFeeIncrease)]
